+++ ServerMonitor/Services/RequestServices/SocketRequest.cs
@@ -52,7 +52,7 @@ namespace ServerMonitor.Services.RequestServices
                                                                  remoteIp, s.ElapsedMilliseconds));
                     #region 修改返回数据
                     Status = "1000";
-                    TimeCost = (short)s.ElapsedMilliseconds;
+                    TimeCost = (int)s.ElapsedMilliseconds;
                     protocolInfo = String.Format("Success, remote server contacted at IP address {0},and the connecting work cost {1} millsseconds!", remoteIp, s.ElapsedMilliseconds);
                     #endregion
                     // 释放连接
@@ -80,7 +80,7 @@ namespace ServerMonitor.Services.RequestServices
                 {
                     #region 修改返回数据
                     Status = "1001";
-                    TimeCost = (short)(OverTime * 1.5);
+                    TimeCost = (int)(OverTime * 1.5);
                     protocolInfo = String.Format("Success, remote server contacted at IP address {0},and the connecting work cost {1} millsseconds!", targetEndPoint.Address, s.ElapsedMilliseconds);
                     #endregion
                     Debug.WriteLine("Error: No such host is known");
@@ -102,7 +102,7 @@ namespace ServerMonitor.Services.RequestServices
                 {
                     #region 修改返回数据
                     Status = "1001";
-                    TimeCost = (short)(OverTime * 1.5);
+                    TimeCost = (int)(OverTime * 1.5);
                     protocolInfo = "Error: Timeout when connecting (check hostname and port)";
                     #endregion
                     Debug.WriteLine("Error: Timeout when connecting (check hostname and port)");
